[ADD] open_academy: New session model creation T#59063
Added session model, a session is an occurrence of a course taught at a given time
 for a given audience.

Created relation between Course and Session.
A session is related to a course, and to a set of attendees (res.partner).

A course has a responsible (res.users).

To have an instructor to relate to certain courses, it was modified the existing Partner model,
as to add an instructor field.

All required views were updated as well.